Refactor Product dialog and screen: unified.
